% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_algorithm.R
\name{calc_matrix_D_best}
\alias{calc_matrix_D_best}
\title{Title}
\usage{
calc_matrix_D_best(data, value_col, instrument_sensitivity = 20)
}
\arguments{
\item{value_col}{Column that contains observed values for creation of the
deconvolution matrix.}

\item{exp_met}{Experimental metadata.}
}
\value{

}
\description{
Title
}
