% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate-manipulation.R
\name{create_blank_plate}
\alias{create_blank_plate}
\title{Creates blank plate with given columns and rows}
\usage{
create_blank_plate(n_cols, n_rows)
}
\arguments{
\item{n_cols}{Number of columns.}

\item{n_rows}{Number of rows.}
}
\value{

}
\description{
Creates blank plate with given columns and rows
}
