% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon_functions.R
\name{create_extended_tibble}
\alias{create_extended_tibble}
\title{Create Extended dataframe from Calibration Plate.}
\usage{
create_extended_tibble(
  data,
  calibrate_row = 5,
  calibrate_col = 5,
  lum_bg_ratio_mean,
  lum_bg_ratio_sd
)
}
\arguments{
\item{lum_bg_ratio_sd}{}
}
\value{

}
\description{
Create Extended dataframe from Calibration Plate.
}
