% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_functions.R
\name{.decon_vec}
\alias{.decon_vec}
\title{Deconvolute a Single Vector}
\usage{
.decon_vec(vec, mat_decon)
}
\arguments{
\item{vec}{Numeric vector, representing the wells.}

\item{mat_decon}{Deconvolution matrix created through \code{.mat_calc_deconvolution()}.}
}
\value{
a numeric vector, the same length as \code{vec}.
}
\description{
Deconvolute a Single Vector
}
