% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_functions.R
\name{.deconvolute_multi_frame_matrix}
\alias{.deconvolute_multi_frame_matrix}
\title{Deconvolute a Multi-Frame Matrix}
\usage{
.deconvolute_multi_frame_matrix(mat_frames, mat_decon)
}
\arguments{
\item{mat_frames}{A multi-frame matrix, with each time point a row, and each
column a well.}

\item{mat_decon}{A deconvolution matrix created through \code{.mat_calc_deconvolution()}.}
}
\value{
A deconvoluted multi-frame matrix.
}
\description{
Deconvolute a Multi-Frame Matrix
}
