% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_functions.R
\name{.mat_calc_bleed}
\alias{.mat_calc_bleed}
\title{Calculate Bleed Through Matrix}
\usage{
.mat_calc_bleed(
  mat,
  ref_row,
  ref_col,
  b_noise = 10,
  relative = TRUE,
  .f = mean
)
}
\arguments{
\item{mat}{A matrix to expand and calculate relative bleed-through based on a
reference well.}

\item{ref_row}{Reference well row number.}

\item{ref_col}{Reference well column number.}

\item{b_noise}{Calculated background noise to fill the bleed-through matrix.}

\item{relative}{TRUE / FALSE whether to return relative values.}

\item{.f}{Function to apply to the expanded matrix, defaults to mean.}
}
\value{
a matrix, that is a bleed-through matrix calculated from the given
matrix.
}
\description{
Expand and calculate a bleed-through matrix from the given matrix around the
given reference well.
}
