% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_functions.R
\name{.mat_calc_deconvolution}
\alias{.mat_calc_deconvolution}
\title{Make Deconvolution Matrix From Bleed-Through Matrix}
\usage{
.mat_calc_deconvolution(mat)
}
\arguments{
\item{mat}{Bleed-through matrix from \code{.mat_calc_bleed()}}
}
\value{
a matrix, ready for deconvolution.
}
\description{
Make Deconvolution Matrix From Bleed-Through Matrix
}
