% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_functions.R
\name{.multi_frame_matrix_from_df}
\alias{.multi_frame_matrix_from_df}
\title{Create a multi-frame matrix from a dataframe.}
\usage{
.multi_frame_matrix_from_df(
  data,
  value,
  time = "time",
  well = "well",
  arrange = FALSE
)
}
\arguments{
\item{data}{A dataframe with columns for the value, time and well ID.}

\item{value}{Name of the column which contains the values for the matrix.}

\item{time}{Name of the column which defines the time points for the frames
(rows) of the matrix.}

\item{well}{Name of the column which contains the well IDs of the samples.}

\item{arrange}{Logical, whether to return a dataframe arranged by time and
well.}
}
\value{
a matrix, with
}
\description{
Creats a matrix with each row being a time point in a multi-frame experiment,
and each column is a single well. The columns are concatenated by row
(ordered A1, A2, ... B1, B2, ...).
}
