% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_functions.R
\name{.multi_frame_matrix_to_vec}
\alias{.multi_frame_matrix_to_vec}
\title{Convert a Multi-Frame Matrix to a Vector}
\usage{
.multi_frame_matrix_to_vec(mat, rowwise = TRUE)
}
\arguments{
\item{mat}{A multi-frame matrix.}

\item{rowwise}{Logical, whether to concatenate the rows (default, TRUE) or
the columns.}
}
\value{
A vector, with values concatenated from the matrix.
}
\description{
Takes a multi-frame matrix and converts to a vector, with each frame being
concatenated end-to-end. Each row / frame of the matrix is a time point, with
each column representing a well (ordered A1, A2, ... B1, B2, ...).
}
