% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_R_functions.R
\name{frames_to_matrix}
\alias{frames_to_matrix}
\title{Title}
\usage{
frames_to_matrix(data, value_col = "lum", time_col = "cycle_nr")
}
\arguments{
\item{time_col}{}
}
\value{

}
\description{
Title
}
