% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate-manipulation.R
\name{matrix_from_tibble}
\alias{matrix_from_tibble}
\title{Creates a matrix from a tibble with row and col columns.}
\usage{
matrix_from_tibble(data, value)
}
\arguments{
\item{data}{Tibble to turn into matrix.}

\item{value}{Column that becomes the values.}
}
\value{

}
\description{
Creates a matrix from a tibble with row and col columns.
}
