% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{num_pad}
\alias{num_pad}
\title{Pad Out Single Digit Numbers}
\usage{
num_pad(x, width = 2)
}
\arguments{
\item{x}{Number to pad out}

\item{width}{How many characters to pad out to.}
}
\value{

}
\description{
Pads out the input, to the left, with zeros, to the given width.
}
