% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{num_to_let}
\alias{num_to_let}
\title{Convert Number to Letter}
\usage{
num_to_let(x)
}
\arguments{
\item{x}{Number to convert to corresponding capital letter.}
}
\value{

}
\description{
Convert Number to Letter
}
