% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{overlay_plate}
\alias{overlay_plate}
\title{Adds plate overlay onto well plot.}
\usage{
overlay_plate(colour = "gray10", size = 1, fill = "NA")
}
\arguments{
\item{colour}{Colour of overlay.}

\item{size}{Size of overlay line.}

\item{fill}{Fill of overlay (default NA).}
}
\value{

}
\description{
Adds plate overlay onto well plot.
}
