% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_wells}
\alias{plot_wells}
\title{Plots Plate of Data}
\usage{
plot_wells(data, fill, log10_fill = TRUE)
}
\arguments{
\item{data}{Data frame with row, col, and column for fill data.}

\item{fill}{Column name for fill data.}

\item{log10_fill}{Logical, whether or not to log10 transform fill data.}
}
\value{

}
\description{
Plots Plate of Data
}
