% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_wells_time}
\alias{plot_wells_time}
\title{Plot Well Values}
\usage{
plot_wells_time(data)
}
\arguments{
\item{data}{Tibble containing \code{cycle_nr}, \code{lum}, and \code{adjusted} columns.}
}
\value{

}
\description{
Compare plotted well values over time.
}
