% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{predict_bleedthrough}
\alias{predict_bleedthrough}
\title{Predict Background Bleedthrough}
\usage{
predict_bleedthrough(dis, height = 3, well_spacing = 25)
}
\arguments{
\item{dis}{Distance between wells.}

\item{height}{Height of the sensor.}

\item{well_spacing}{Spacing of the wells.}
}
\value{

}
\description{
Predict Background Bleedthrough
}
