% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon_functions.R
\name{random_extended_matrix}
\alias{random_extended_matrix}
\title{Create Randomised Extended Matrix for Optimisation}
\usage{
random_extended_matrix(data)
}
\arguments{
\item{data}{Extended tibble from \code{create_extended_tibble()}.}
}
\value{

}
\description{
Create Randomised Extended Matrix for Optimisation
}
