% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-plate.R
\name{read_plate}
\alias{read_plate}
\title{Read luminescent values from a plate.}
\usage{
read_plate(path)
}
\arguments{
\item{path}{Path to excel file with luminescent values.}
}
\value{

}
\description{
Read luminescent values from a plate.
}
