% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_functions.R
\name{rl_adjust_plate}
\alias{rl_adjust_plate}
\title{Adjust Experimental Luminescent Data}
\usage{
rl_adjust_plate(data, value, mat_decon, time = "time", well = "well")
}
\arguments{
\item{data}{A data frame that contains the experimental data.}

\item{value}{The name of the column containing the values (i.e. 'lum').}

\item{mat_decon}{A deconvolution matrix created through
\code{rl_calc_decon_matrix()}}

\item{time}{The name of the column containing the time values (i.e.
'time')}
}
\value{
A dataframe with the specified column having been deconvoluted, using
the supplied deconvolution matrix.
}
\description{
Using a deconvolution matrix, created through \code{rl_calc_decon_matrix()},
adjust the values in the \code{col_value} column to take into account
bleed-through from surrounding wells.
}
