% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_functions.R
\name{rl_calc_decon_matrix}
\alias{rl_calc_decon_matrix}
\title{Calculate a Deconvolution Matrix}
\usage{
rl_calc_decon_matrix(
  data,
  value,
  b_noise,
  time = "time",
  ref_well = "I05",
  well = "well"
)
}
\arguments{
\item{data}{A data frame that contains the data of the calibration plate.}

\item{value}{Name of the column containing the luminescent values.}

\item{b_noise}{The value of the background noise, which is the average signal
for the background wells that are far away from the reference well.}

\item{time}{Name of the column with the time values.}

\item{ref_well}{The well ID of the reference well (i.e. 'E05', 'I12")}

\item{well}{Name of the column with the well ID values.}
}
\value{
a deconvolution matrix, for use in \code{rl_adjust_plate()}
}
\description{
Use the data from a calibration plate, where the plate is empty except for
a single well with a luminescent signal, to create a deconvolution matrix
that can be used to adjust other experimental results.
}
\details{
The deconvolution matrix will be unique for each plate type and plate-reader,
so a matrix should be calculated for each combination of plate and plate
reader, but once this is calculated, it can be re-used to adjust future
experimental results.
}
\examples{

fl <- system.file(
  "extdata",
  "calibrate_tecan",
  "calTecan1.xlsx",
  package = "reluxr"
)

dat <- plate_read_tecan(fl)

dat

mat_d <- dat |>
  dplyr::filter(signal != "OD600") |>
  dplyr::filter(time_s > 500) |>
  rl_calc_decon_matrix(value, time_s, ref_well = "E05", b_noise = 30)

image(log10(mat_d))
}
