% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_functions.R
\name{rl_decon_frames}
\alias{rl_decon_frames}
\title{Deconvolute a Multi-Frame Matrix}
\usage{
rl_decon_frames(mat_frames, decon_mat)
}
\arguments{
\item{mat_frames}{A multi-frame matrix, with each time point a row, and each
column a well.}

\item{decon_mat}{A deconvolution matrix created through \code{rl_mat_decon()}.}
}
\value{
A deconvoluted multi-frame matrix.
}
\description{
Deconvolute a Multi-Frame Matrix
}
