% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_bleed.R
\name{rl_decon_vec}
\alias{rl_decon_vec}
\title{Deconvolute a Single Vector}
\usage{
rl_decon_vec(vec, decon_mat)
}
\arguments{
\item{vec}{Numeric vector, representing the wells.}

\item{decon_mat}{Deconvolution matrix created through \code{rl_mat_decon()}.}
}
\value{
a numeric vector, the same length as \code{vec}.
}
\description{
Deconvolute a Single Vector
}
