% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_functions.R
\name{rl_df_decon_best}
\alias{rl_df_decon_best}
\title{Title}
\usage{
rl_df_decon_best(
  data,
  values_col,
  time_col,
  ref_well = "I05",
  b_noise,
  well_col = "well"
)
}
\arguments{
\item{b_noise}{}
}
\value{

}
\description{
Title
}
