% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_bleed.R
\name{rl_mat_bleed}
\alias{rl_mat_bleed}
\title{Calculate Bleed Through Matrix}
\usage{
rl_mat_bleed(mat, ref_row, ref_col, b_noise = 10, relative = TRUE, .f = mean)
}
\arguments{
\item{mat}{A matrix to expand and calculate relative bleed-through based on a
reference well.}

\item{ref_row}{Reference row number.}

\item{ref_col}{Reference column number.}

\item{b_noise}{Calculated background noise to fill the bleed-through matrix.}

\item{relative}{TRUE / FALSE whether to return relative values.}

\item{.f}{Function to apply to the expanded matrix, defaults to mean.}
}
\value{
a matrix, that is a bleed-through matrix calculated from the given
matrix.
}
\description{
Calculate Bleed Through Matrix
}
