% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_bleed.R
\name{rl_mat_decon}
\alias{rl_mat_decon}
\title{Make Deconvolution Matrix From Bleed-Through}
\usage{
rl_mat_decon(mat)
}
\arguments{
\item{mat}{Bleed-through matrix from \code{rl_mat_bleed()}}
}
\value{
a matrix, ready for deconvolution.
}
\description{
Make Deconvolution Matrix From Bleed-Through
}
