% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_bleed.R
\name{rl_mat_decon_best}
\alias{rl_mat_decon_best}
\title{Calculate the Optimal Deconvolution Matrix}
\usage{
rl_mat_decon_best(mat, ref_row, ref_col, b_noise = 20)
}
\arguments{
\item{mat}{A multi-frame matrix, where each row is a frame and each column is a well.}

\item{ref_row}{Row number for the reference well.}

\item{ref_col}{Column number for the reference well.}

\item{b_noise}{Value for the background noise.}
}
\value{
a matrix, the optimised deconvolution matrix.
}
\description{
Calculate the Optimal Deconvolution Matrix
}
