% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rl_plot_plate}
\alias{rl_plot_plate}
\title{Plot a Plate-Layout of the Value Column}
\usage{
rl_plot_plate(data, value, well = "well", trans = log10)
}
\arguments{
\item{data}{Dataframe with the value column and a column specifying the well
ID.}

\item{value}{Name of the column containing the value information to be
displayed.}

\item{well}{Name of the column with the well ID information for formatting
the plate layout.}

\item{trans}{Name of the transformation to apply to the data. Defaults to
\code{log10}.}
}
\value{
a \code{ggplot2::ggplot()} plot.
}
\description{
Plot a Plate-Layout of the Value Column
}
\examples{
library(ggplot2)
library(dplyr)

fl <- system.file(
  "extdata",
  "calibrate_tecan",
  "calTecan1.xlsx",
  package = "reluxr"
)

dat <- plate_read_tecan(fl)

mat_d_best <- dat |>
  dplyr::filter(signal == "LUMI") |>
  dplyr::filter(time_s > 500) |>
  rl_calc_decon_matrix("value", "time_s", ref_well = "E05", b_noise = 30)

dat |>
  ungroup() |>
  dplyr::filter(signal == "LUMI") |>
  dplyr::filter(time_s > 500) |>
  rl_adjust_plate(value, mat_d_best, time = time_s) |>
  summarise(value = mean(value), .by = well) |>
  rl_plot_plate(value, trans = log10) +
  scale_fill_viridis_c(
    limits = c(1, NA)
  )

}
