% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate-manipulation.R
\name{tibble_from_vec}
\alias{tibble_from_vec}
\title{Turns a vector into a palte-based tibble.}
\usage{
tibble_from_vec(vec, n_row = 8, n_col = 12)
}
\arguments{
\item{vec}{Vector of values.}
}
\value{

}
\description{
Turns a vector into a palte-based tibble.
}
