% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate-manipulation.R
\name{tibble_to_matrix}
\alias{tibble_to_matrix}
\title{Creates a matrix from a tibble with row and col columns.}
\usage{
tibble_to_matrix(data, value)
}
\arguments{
\item{data}{Tibble to turn into matrix.}

\item{value}{Column that becomes the values.}
}
\value{

}
\description{
Creates a matrix from a tibble with row and col columns.
}
