% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_R_functions.R
\name{translate_df_wells}
\alias{translate_df_wells}
\title{Title}
\usage{
translate_df_wells(data, calibrate_row = 5, calibrate_col = 5)
}
\arguments{
\item{calibrate_col}{}
}
\value{

}
\description{
Title
}
