% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_algorithm.R
\name{update_frames}
\alias{update_frames}
\title{Update Frames Matrix}
\usage{
update_frames(old_mat, new_mat, cutoff = 0)
}
\arguments{
\item{old_mat}{Frames matrix that is updated.}

\item{new_mat}{Frames matrix that contains deconvoluted values.}

\item{cutoff}{Cutoff for comparison. Default = 0.}
}
\value{
Frames matrix
}
\description{
Updates the old_mat with values from new_mat if they are below the old_mat or
less than or equal to the cutoff (default 0).
}
