% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate-manipulation.R
\name{vec_to_tibble}
\alias{vec_to_tibble}
\title{Turns a vector into a palte-based tibble.}
\usage{
vec_to_tibble(vec, n_row = 8, n_col = 12)
}
\arguments{
\item{vec}{Vector of values.}
}
\value{

}
\description{
Turns a vector into a palte-based tibble.
}
