% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{well_dis}
\alias{well_dis}
\title{Calculate Distance Between Wells}
\usage{
well_dis(row, col, calibrate_row = 5, calibrate_col = 5)
}
\arguments{
\item{row}{Vector of row numbers.}

\item{col}{Vector of column numbers.}

\item{calibrate_row}{Single row number to measure from.}

\item{calibrate_col}{Single column number to measure from.}
}
\value{

}
\description{
Calculates distance between row and column positions, relative to a calibration well given by \code{calibrate_row} and \code{calibrate_col}.
}
