% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{well_format}
\alias{well_format}
\title{Take a well ID and format it evrbosely.}
\usage{
well_format(x)
}
\arguments{
\item{x}{Vector of well IDs}
}
\value{
Vector of formatted well IDs as strings.
}
\description{
Take a well ID and format it evrbosely.
}
\examples{
well_format(c("a05", "a5", "A5", "C1"))
}
