% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{well_join}
\alias{well_join}
\title{Join a row and column into a well ID.}
\usage{
well_join(row, col, num_width = 2)
}
\arguments{
\item{num_width}{}
}
\value{

}
\description{
Joins a column number and a row letter or number into a well ID. i.e. joins
"A" and "1" to "A01" and joins "3 and "10" to "C10".
}
