% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_algorithm.R
\name{well_to_index}
\alias{well_to_index}
\title{Converts a well ID string into a Vector index}
\usage{
well_to_index(well, n_wells = 96)
}
\arguments{
\item{well}{Well ID string to convert to index.}

\item{n_wells}{Number of wells in the plate.}
}
\value{

}
\description{
For indexing vectors and frames matrices, converts a string well ID ("B05")
into an integer for indexing into the columns of a frames matrix or a vector,
given the number of wells in the plate.
}
