% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{well_to_rownum}
\alias{well_to_rownum}
\title{Convert Well ID to Row Number}
\usage{
well_to_rownum(x)
}
\arguments{
\item{x}{Well ID as a string.}
}
\value{
Numeric row number.
}
\description{
Convert Well ID to Row Number
}
