% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon_functions.R
\name{calc_bleed_df}
\alias{calc_bleed_df}
\title{Calculate Bleed-Through Dataframe}
\usage{
calc_bleed_df(data, time_cutoff = 30, calibrate_row = 5, calibrate_col = 5)
}
\arguments{
\item{calibrate_col}{}
}
\value{

}
\description{
Calculate Bleed-Through Dataframe
}
