% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_algorithm.R
\name{calc_matrix_D_best}
\alias{calc_matrix_D_best}
\title{Title}
\usage{
calc_matrix_D_best(
  data,
  value_col,
  col_time,
  col_value,
  calibration_well = "E05",
  instrument_sensitivity = 20,
  time_cutoff = 50
)
}
\arguments{
\item{value_col}{Column that contains observed values for creation of the}

\item{time_cutoff}{deconvolution matrix.}
}
\value{

}
\description{
Title
}
