% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_functions.R
\name{rl_adjust_plate}
\alias{rl_adjust_plate}
\title{Adjust Experimental Luminescent Data}
\usage{
rl_adjust_plate(data, col_value, col_time, mat_decon)
}
\arguments{
\item{data}{A data frame that contains the experimental data.}

\item{col_value}{The name of the column containing the values (i.e. 'lum').}

\item{col_time}{The name of the column containing the time values (i.e.
'time')}

\item{mat_decon}{A deconvolution matrix created through
\code{rl_calc_decon_matrix()}}
}
\value{

}
\description{
Using a deconvolution matrix, created through \code{rl_calc_decon_matrix()},
adjust the values in the \code{col_value} column to take into account
bleed-through from surrounding wells.
}
